requirements.txt
python-telegram-bot
google-search-results